Create control component
