+++ src/containers/MissionControl.js
@@ -1,29 +1,22 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { Button, Grid, Header } from 'semantic-ui-react';
+import { Grid, Header } from 'semantic-ui-react';
 import { hot } from 'react-hot-loader';
 import PropTypes from 'prop-types';
 
-import {
-  changeExecutionState as actionChangeExecutionState,
-  EXECUTION_RUN,
-  EXECUTION_STEP,
-  EXECUTION_STOP,
-  EXECUTION_RESET,
-} from '@/actions/code';
 import Console from '@/components/Console';
+import Control from '@/components/Control';
 import Workspace from '@/components/Workspace';
 
 const mapStateToProps = ({ code }) => ({ code });
-const mapDispatchToProps = dispatch => ({
-  changeExecutionState: state => dispatch(actionChangeExecutionState(state)),
-});
 
-const MissionControl = ({ code, changeExecutionState }) => (
+const MissionControl = ({ code }) => (
   <Grid columns={16} divided>
     <Grid.Row>
       <Grid.Column width={6}>
-        <Workspace />
+        <Grid.Row>
+          <Workspace />
+        </Grid.Row>
       </Grid.Column>
       <Grid.Column width={6}>
         <Grid.Row>
@@ -35,18 +28,7 @@ const MissionControl = ({ code, changeExecutionState }) => (
           </pre>
         </Grid.Row>
         <Grid.Row>
-          <Button onClick={() => changeExecutionState(EXECUTION_STEP)}>
-            Step
-          </Button>
-          <Button onClick={() => changeExecutionState(EXECUTION_RUN)}>
-            Run
-          </Button>
-          <Button onClick={() => changeExecutionState(EXECUTION_STOP)}>
-            Stop
-          </Button>
-          <Button onClick={() => changeExecutionState(EXECUTION_RESET)}>
-            Reset
-          </Button>
+          <Control />
         </Grid.Row>
       </Grid.Column>
       <Grid.Column width={4}>
@@ -60,7 +42,6 @@ MissionControl.propTypes = {
   code: PropTypes.shape({
     jsCode: PropTypes.string,
   }).isRequired,
-  changeExecutionState: PropTypes.func.isRequired,
 };
 
-export default hot(module)(connect(mapStateToProps, mapDispatchToProps)(MissionControl));
+export default hot(module)(connect(mapStateToProps)(MissionControl));
